feat(#211): Spam CreateAsync sub-step + ValidatePathsSet per-folder probe (Phase 3.5)
Isolates which sub-step of the Spam engine init consumes the proven ~113s STA
freeze. Emits [spam-init] lines timing each CreateAsync sub-step (ValidatePathsSet,
ValidateSpamClassifier, InitAsync model load) and each ValidatePathsSet folder COM
read (JunkCertain/JunkPotential/Inbox), so the next slow cold-start capture names
the exact blocking COM call.

- Brought SpamBayes.cs from 705 -> 446 lines via three behavior-preserving partial
  files (Conditions/Actions/Classify), fixing a pre-existing 500-line violation
- Pure formatting in new coverable SpamInitTimingProbe; 13 deterministic MSTests;
  new code 100%; UtilitiesCS 87.17% (>=80% floor held); no repo regression
- Behavior-preserving; Stopwatch only; no banned APIs; model deserialize +
  PreserveReferencesHandling + StoreWrapper instrumentation untouched
- Maintainer slow cold-start capture instructions + placeholder under evidence/other/
