+++ UtilitiesCS/UtilitiesCS.csproj
@@ -654,6 +654,10 @@
     <Compile Include="EmailIntelligence\EmailParsingSorting\ImageStripper.cs" />
     <Compile Include="OutlookObjects\Attachment\AttachmentSerializable.cs" />
     <Compile Include="EmailIntelligence\ClassifierGroups\SpamBayes\SpamBayes.cs" />
+    <Compile Include="EmailIntelligence\ClassifierGroups\SpamBayes\SpamBayes.Conditions.cs" />
+    <Compile Include="EmailIntelligence\ClassifierGroups\SpamBayes\SpamBayes.Actions.cs" />
+    <Compile Include="EmailIntelligence\ClassifierGroups\SpamBayes\SpamBayes.Classify.cs" />
+    <Compile Include="EmailIntelligence\ClassifierGroups\SpamBayes\SpamInitTimingProbe.cs" />
     <Compile Include="EmailIntelligence\ClassifierGroups\TristateEngine.cs" />
     <Compile Include="EmailIntelligence\ClassifierGroups\Triage\Triage.cs" />
     <Compile Include="EmailIntelligence\ClassifierGroups\ManagerAsyncLazy.cs" />
