+++ docs/features/active/2026-06-22-outlook-startup-intelconfig-continuation-stall-211/evidence/baseline/baseline-csharpier-2026-06-24T15-10.md
@@ -1,0 +1,11 @@
+# Baseline — CSharpier Check (issue #211)
+
+Timestamp: 2026-06-24T15-10
+
+Command: `dotnet tool run csharpier check .` (CSharpier v1.2.6 local tool; v1 subcommand syntax)
+
+EXIT_CODE: 0
+
+Output Summary:
+- Result: PASS. `Checked 1095 files in 2451ms.` No formatting violations reported.
+- Baseline repository formatting is clean prior to any edit.
