+++ docs/features/active/2026-06-22-outlook-startup-intelconfig-continuation-stall-211/evidence/qa-gates/final-qc-analyzers-2026-06-24T15-10.md
@@ -1,0 +1,14 @@
+# Final QC — Analyzer Build (issue #211)
+
+Timestamp: 2026-06-24T15-10
+
+Command: `msbuild TaskMaster.sln /t:Build /p:Configuration=Debug /p:Platform="Any CPU" /p:EnableNETAnalyzers=true /p:EnforceCodeStyleInBuild=true`
+(executed via git-bash with dash-switches)
+
+EXIT_CODE: 0
+
+Output Summary:
+- Result: `Build succeeded. 0 Warning(s) 0 Error(s)`.
+- No analyzer diagnostics introduced by the partial-class extractions, the `SpamInitTimingProbe`
+  helper, or the `SpamBayes.CreateAsync`/`ValidatePathsSet` instrumentation.
+- No files changed by this step; loop continues.
