+++ docs/features/active/2026-06-22-outlook-startup-intelconfig-continuation-stall-211/evidence/qa-gates/final-qc-filesize-2026-06-24T15-10.md
@@ -1,0 +1,25 @@
+# Final QC — File-Size Verification (CONSTRAINT-A) (issue #211)
+
+Timestamp: 2026-06-24T15-10
+
+Command: `wc -l <file>` for each touched/new file.
+
+EXIT_CODE: 0
+
+## Post-change line counts
+
+| File | Lines | <= 500? |
+|---|---|---|
+| `UtilitiesCS/EmailIntelligence/ClassifierGroups/SpamBayes/SpamBayes.cs` | 446 | PASS |
+| `UtilitiesCS/EmailIntelligence/ClassifierGroups/SpamBayes/SpamBayes.Conditions.cs` | 100 | PASS |
+| `UtilitiesCS/EmailIntelligence/ClassifierGroups/SpamBayes/SpamBayes.Actions.cs` | 117 | PASS |
+| `UtilitiesCS/EmailIntelligence/ClassifierGroups/SpamBayes/SpamBayes.Classify.cs` | 121 | PASS |
+| `UtilitiesCS/EmailIntelligence/ClassifierGroups/SpamBayes/SpamInitTimingProbe.cs` | 81 | PASS |
+| `UtilitiesCS.Test/EmailIntelligence/SpamInitTimingProbeTests.cs` | 214 | PASS |
+
+## CONSTRAINT-A verdict
+
+PASS. `SpamBayes.cs` was reduced from 705 lines (baseline, 205 over the 500 ceiling) to 446 lines
+via the three partial-class extractions, then the Phase 3 Stopwatch instrumentation added net lines
+back while remaining at 446 — well within the 500 ceiling. Every touched and new file is <= 500
+lines. CONSTRAINT-A is satisfied.
