+++ docs/features/active/2026-06-22-outlook-startup-intelconfig-continuation-stall-211/evidence/qa-gates/final-qc-nullable-2026-06-24T15-10.md
@@ -1,0 +1,15 @@
+# Final QC — Nullable / TreatWarningsAsErrors Build (issue #211)
+
+Timestamp: 2026-06-24T15-10
+
+Command: `msbuild TaskMaster.sln /t:Build /p:Configuration=Debug /p:Platform="Any CPU" /p:Nullable=enable /p:TreatWarningsAsErrors=true`
+(executed via git-bash with dash-switches)
+
+EXIT_CODE: 0
+
+Output Summary:
+- Result: `Build succeeded. 0 Warning(s) 0 Error(s)`.
+- No nullable-flow warnings (promoted to errors under TWAE) introduced by the touched/new code.
+- The `SpamInitTimingProbe` null-guards (`ArgumentNullException` on null `emit`/`step`) and the
+  nullable-clean instrumentation in `SpamBayes` pass the protected nullable gate.
+- No files changed by this step; loop continues.
