+++ docs/features/active/2026-06-22-outlook-startup-intelconfig-continuation-stall-211/spec.md
@@ -185,6 +185,8 @@ Eliminate the multi-minute Outlook startup latency. The current evidence localiz
 
 - [x] AC16: Per-store filter attribution probe (diagnosis-only). For each store enumerated by `StoresWrapper.GetFilteredStores()`/`ShouldIncludeStore`, exactly one structured `[store-filter]` line is emitted via the existing `log4net` logger capturing store DisplayName (guarded), `Stopwatch` ms to read `ExchangeStoreType`, `Stopwatch` ms to read `FilePath`, the include/exclude boolean, and the matched rule (`PublicFolder`/`NameContains`/`GwsoFilePath`/`FilePathContains`/`Included`). On the synchronous `Init()` path a `GetFilteredStores`-level `[store-filter]` summary (count + total ms) is emitted; the async `RewireOlObjectsAsync` path is unchanged. Behavior-preserving (identical filter result, included set, order, predicate short-circuit semantics); `Stopwatch` only; no banned APIs; net48; all touched files <= 500 lines; `PreserveReferencesHandling.All` and existing instrumentation tags untouched. Pure decision + line formatter live in the coverable `StoreFilterAttribution` helper and are covered by deterministic MSTest (no live COM/timer/filesystem/network; no temporary files). — AUTOMATED PORTION VERIFIED: implementation in `UtilitiesCS/OutlookObjects/Store/StoreFilterAttribution.cs` (new, 100% new-code coverage) and `StoresWrapper.cs`; 13 deterministic tests pass; full C# toolchain passes in order (CSharpier -> analyzers -> nullable/TWAE -> MSTest with coverage); no repo-wide regression; all touched files <= 500 lines. Evidence under `evidence/qa-gates/*-2026-06-24T14-30.md` and `evidence/baseline/*-2026-06-24T14-30.md`.
 - [~] AC16 (runtime, maintainer): a non-debugger cold-start capture records the `[store-filter]` lines during a slow startup so the largest `FilePath`/`ExchangeStoreType` read time and the Gmail store's include/exclude decision are identifiable. — MAINTAINER-GATED (runtime, not CI-automatable). Capture instructions: `evidence/other/coldstart-store-filter-capture-instructions-2026-06-24T14-30.md`; placeholder: `evidence/other/runtime-capture-store-filter-PLACEHOLDER.md`. Pending maintainer capture.
+- [x] AC17: Per-sub-step + per-folder SpamBayes-init attribution probe (diagnosis-only). `SpamBayes.CreateAsync` emits exactly one structured `[spam-init]` line per sub-step (`ValidatePathsSet`, `ValidateSpamClassifier`, `InitAsync(modelLoad)`) via the existing `log4net` logger with `Stopwatch` F1 ms, and `ValidatePathsSet()` emits one `[spam-init]` line per COM folder access (`ValidatePathsSet.JunkCertain`, `ValidatePathsSet.JunkPotential`, `ValidatePathsSet.Inbox`) so the exact blocking sub-step and folder are identifiable relative to the proven ~113 s freeze. Behavior-preserving (identical validation result, identical exceptions, identical engine returned; model deserialize and `PreserveReferencesHandling.All` untouched; existing `[Startup timing]`/`[engine-init]` lines untouched). `Stopwatch` only; no banned APIs (`DateTime.Now`/`UtcNow`, `Random.Shared`, `Thread.Sleep`, `Task.Delay`); net48. The pure line formatter lives in the coverable `SpamInitTimingProbe` helper (NOT `[ExcludeFromCodeCoverage]`), covered by deterministic MSTest (MSTest + Moq + FluentAssertions) with a list-capturing sink and no live COM/timer/filesystem/network and no temporary files; new code `>= 90%` coverage with no repository-wide regression. All touched and new files `<= 500` lines (CONSTRAINT-A: `SpamBayes.cs` reduced to `<= 500`, expected ~402, via the `SpamBayes.Conditions.cs`, `SpamBayes.Actions.cs`, and `SpamBayes.Classify.cs` partial-class extractions). Full C# toolchain passes in order (CSharpier -> analyzers -> nullable/TWAE -> MSTest with coverage, `TestCategory!=LiveOutlook`). Runtime portion (maintainer): a non-debugger cold-start capture records the six `[spam-init]` lines during a slow startup; capture instructions and placeholder under `evidence/other/`. — AUTOMATED PORTION VERIFIED: implementation in `UtilitiesCS/EmailIntelligence/ClassifierGroups/SpamBayes/SpamInitTimingProbe.cs` (new, 100% new-code coverage), the three partial-class extractions (`SpamBayes.Conditions.cs`/`SpamBayes.Actions.cs`/`SpamBayes.Classify.cs`), and the `SpamBayes.CreateAsync`/`ValidatePathsSet` instrumentation; 13 deterministic tests pass; full C# toolchain passes in order; no repo-wide regression (60.43% -> 60.47%); `SpamBayes.cs` reduced 705 -> 446 lines; all six touched/new files <= 500. Evidence under `evidence/qa-gates/*-2026-06-24T15-10.md` and `evidence/baseline/*-2026-06-24T15-10.md`.
+- [~] AC17 (runtime, maintainer): a non-debugger cold-start capture records the six `[spam-init]` lines during a slow startup so the dominant sub-step (`ValidatePathsSet` / `ValidateSpamClassifier` / `InitAsync(modelLoad)`) and dominant folder (`JunkCertain` / `JunkPotential` / `Inbox`) are identifiable relative to the proven ~113 s freeze. — MAINTAINER-GATED (runtime, not CI-automatable). Capture instructions: `evidence/other/coldstart-spam-init-capture-instructions-2026-06-24T15-10.md`; placeholder: `evidence/other/runtime-capture-spam-init-PLACEHOLDER.md`. Pending maintainer capture.
 
 > Separate finding (new scope, not the latency): the Spam/Triage/Actionable inbox custom-column fields remain empty for new emails (`ProcessNewInboxItemsAsync` elapsedMs=0). The classification path `AppEvents.OlInboxItems_ItemAdd -> ProcessMailItemAsync` is independent of engine construction and was not modified by Phase 3. Tracked as a separate concern; see the cold-start capture file.
